@@ -2,26 +2,25 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace CK.SqlServer
+namespace CK.SqlServer;
+
+/// <summary>
+/// Defines the three states of a transaction.
+/// </summary>
+public enum SqlTransactionStatus
 {
     /// <summary>
-    /// Defines the three states of a transaction.
+    /// The transaction is opened but not yet committed nor rollbacked.
     /// </summary>
-    public enum SqlTransactionStatus
-    {
-        /// <summary>
-        /// The transaction is opened but not yet committed nor rollbacked.
-        /// </summary>
-        Opened,
+    Opened,
 
-        /// <summary>
-        /// The transaction has been committed.
-        /// </summary>
-        Committed,
+    /// <summary>
+    /// The transaction has been committed.
+    /// </summary>
+    Committed,
 
-        /// <summary>
-        /// The transaction has been rollbacked.
-        /// </summary>
-        Rollbacked
-    }
+    /// <summary>
+    /// The transaction has been rollbacked.
+    /// </summary>
+    Rollbacked
 }
