@@ -1,15 +1,14 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CK.SqlServer
+namespace CK.SqlServer;
+
+/// <summary>
+/// Defines a disposable <see cref="ISqlCallContext"/>. 
+/// </summary>
+public interface IDisposableSqlCallContext : ISqlCallContext, IDisposable
 {
-    /// <summary>
-    /// Defines a disposable <see cref="ISqlCallContext"/>. 
-    /// </summary>
-    public interface IDisposableSqlCallContext : ISqlCallContext, IDisposable
-    {
-    }
 }
