@@ -1,12 +1,11 @@
-namespace CodeCake.Abstractions
-{
-    interface ICIPublishWorkflow
-    {
-        /// <summary>
-        /// Pack the solution: it produce the artifacts.
-        /// </summary>
-        void Pack();
+namespace CodeCake.Abstractions;
 
-        ArtifactType ArtifactType { get; }
-    }
+interface ICIPublishWorkflow
+{
+    /// <summary>
+    /// Pack the solution: it produce the artifacts.
+    /// </summary>
+    void Pack();
+
+    ArtifactType ArtifactType { get; }
 }
